moved `pc.state` files to `pc`
