+++ presentation-compiler/src/main/scala/org/alephium/ralph/lsp/pc/PCState.scala
@@ -1,4 +1,4 @@
-package org.alephium.ralph.lsp.pc.state
+package org.alephium.ralph.lsp.pc
 
 import org.alephium.ralph.lsp.pc.workspace.WorkspaceState
 import org.alephium.ralph.lsp.pc.workspace.build.BuildState
